pyproject.toml
[project]
name = "arcframe-engine"
version = "0.1.0"
description = "Event-driven engine architecture for pygame-ce."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pygame-ce>=2.5.7",
]

[project.scripts]
example = "arcframe.example.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/arcframe"]

[tool.uv]
package = true